Fix gcc warnings about signed/unsigned comparison
Reviewed-on: http://codereview.qt.nokia.com/2860

+++ src/plugins/bineditor/bineditor.cpp
@@ -731,7 +731,7 @@ void BinEditor::paintEvent(QPaintEvent *e)
     QString itemString(m_bytesPerLine*3, QLatin1Char(' '));
     QChar *itemStringData = itemString.data();
     char changedString[160] = { false };
-    QTC_ASSERT(m_bytesPerLine < sizeof(changedString), return);
+    QTC_ASSERT((size_t)m_bytesPerLine < sizeof(changedString), return);
     const char *hex = "0123456789abcdef";
 
     painter.setPen(palette().text().color());
